requirements.txt
phidata
python-dotenv
groq
yfinance
duckduckgo-search